requirements.txt
amqp==5.3.1
asgiref==3.8.1
celery==5.4.0
Django==5.1.3
djangorestframework==3.15.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
redis==5.2.0
sqlparse==0.5.2
django-environ==0.11.2
pydantic==2.10.1
